Removed swappable.Handler, multichannelfanout.Handler, fanout.Handler
